+++ Card_Info.py
@@ -1,5 +1,4 @@
 #// auth_ Mohamad Janati
-#// AmirHassan Asvadi ;)
 #// Copyright (c) 2019-2020 Mohamad Janati (freaking stupid, right? :|)
 
 import os
@@ -255,19 +254,19 @@ class StatsSidebar(object):
                 pressed_good = mw.col.db.scalar("select sum(case when ease = 2 then 1 else 0 end) from revlog where cid = ?", c.id)
                 pressed_easy = mw.col.db.scalar("select sum(case when ease = 3 then 1 else 0 end) from revlog where cid = ?", c.id)
                 pressed_all = pressed_again + pressed_good + pressed_easy
-                self.addLine("Again", "{} | {:.0f}%".format(pressed_again, float(pressed_again/pressed_all)*100))
-                self.addLine("Good", "{} | {:.0f}%".format(pressed_good, float(pressed_good/pressed_all)*100))
-                self.addLine("Easy", "{} | {:.0f}%".format(pressed_easy, float(pressed_easy/pressed_all)*100))
+                self.addLine("Again", "{} | {:.0f}%".format(str(pressed_again).rjust(4), float(pressed_again/pressed_all)*100))
+                self.addLine("Good", "{} | {:.0f}%".format(str(pressed_good).rjust(4), float(pressed_good/pressed_all)*100))
+                self.addLine("Easy", "{} | {:.0f}%".format(str(pressed_easy).rjust(4), float(pressed_easy/pressed_all)*100))
             elif self.col.schedVer() == 2:
                 pressed_again = mw.col.db.scalar("select sum(case when ease = 1 then 1 else 0 end) from revlog where cid = ?", c.id)
                 pressed_hard = mw.col.db.scalar("select sum(case when ease = 2 then 1 else 0 end) from revlog where cid = ?", c.id)
                 pressed_good = mw.col.db.scalar("select sum(case when ease = 3 then 1 else 0 end) from revlog where cid = ?", c.id)
                 pressed_easy = mw.col.db.scalar("select sum(case when ease = 4 then 1 else 0 end) from revlog where cid = ?", c.id)
                 pressed_all = pressed_again + pressed_hard + pressed_good + pressed_easy
-                self.addLine("Again", "{} | {:.0f}%".format(pressed_again, float(pressed_again/pressed_all)*100))
-                self.addLine("Hard", "{} | {:.0f}%".format(pressed_hard, float(pressed_hard/pressed_all)*100))
-                self.addLine("Good", "{} | {:.0f}%".format(pressed_good, float(pressed_good/pressed_all)*100))
-                self.addLine("Easy", "{} | {:.0f}%".format(pressed_easy, float(pressed_easy/pressed_all)*100))
+                self.addLine("Again", "{} | {:.0f}%".format(str(pressed_again).rjust(4), float(pressed_again/pressed_all)*100))
+                self.addLine("Hard", "{} | {:.0f}%".format(str(pressed_hard).rjust(4), float(pressed_hard/pressed_all)*100))
+                self.addLine("Good", "{} | {:.0f}%".format(str(pressed_good).rjust(4), float(pressed_good/pressed_all)*100))
+                self.addLine("Easy", "{} | {:.0f}%".format(str(pressed_easy).rjust(4), float(pressed_easy/pressed_all)*100))
             if infobar_reviews:
                 self.addLine("Reviews", "%d" % c.reps)
             (cnt, total) = self.col.db.first("select count(), sum(time)/1000 from revlog where cid = ?", c.id)